pyproject.toml
[project]
name = "mlh-pe-hackathon"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "faker>=33.0",
    "flask>=3.1",
    "peewee>=3.17",
    "prometheus-client>=0.21",
    "pytest>=8.3",
    "pytest-cov>=6.0",
    "redis>=6.0",
    "psycopg2-binary>=2.9",
    "python-dotenv>=1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = [
    "--cov=app",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=xml",
]

[tool.coverage.run]
branch = true
source = ["app"]

[tool.coverage.report]
show_missing = true
skip_covered = true
omit = ["tests/*"]
